# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhuketDeal Real Estate Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  post_id: string;
  post_url?: string | null;
  author_name?: string | null;
  author_profile_url?: string | null;
  post_text?: string | null;
  created_at?: string | null;
  comment_count?: number | null;
  like_count?: number | null;
  screenshot_path?: string | null;
  classification_label?: string | null;
  is_direct_owner?: boolean | null;
  is_agent?: boolean | null;
  is_hot?: boolean | null;
  source_mode?: string | null;
  listing_type?: string | null;
  price_thb?: number | null;
  price_thb_min?: number | null;
  price_thb_max?: number | null;
  price_unit?: string | null;
  price_negotiable?: boolean | null;
  property_type?: string | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  floors?: number | null;
  size_sqm?: number | null;
  land_rai?: number | null;
  location?: string | null;
  district?: string | null;
  phuket_zone?: string | null;
  project_name?: string | null;
  condition?: string | null;
  furnished?: string | null;
  has_pool?: boolean | null;
  has_sea_view?: boolean | null;
  has_garden?: boolean | null;
  has_parking?: boolean | null;
  has_gym?: boolean | null;
  has_security?: boolean | null;
  has_mountain_view?: boolean | null;
  poster_type?: string | null;
  urgency_signals?: string[] | null;
  key_features?: string[] | null;
  decision_summary?: string | null;
  opportunity_score?: string | null;
  confidence?: string | null;
}

export interface StatsOverview {
  total_posts_this_week: number;
  total_posts_all_time: number;
  direct_owner_count: number;
  direct_owner_pct: number;
  for_sale_count: number;
  for_rent_count: number;
  high_opportunity_count: number;
  avg_price_thb?: number | null;
}

export interface DistrictStat {
  district: string;
  count: number;
}

export interface PropertyTypeStat {
  property_type: string;
  count: number;
}

export interface PricePerSqmStat {
  district: string;
  avg_price_per_sqm?: number | null;
  median_price_per_sqm?: number | null;
  min_price_per_sqm?: number | null;
  max_price_per_sqm?: number | null;
  count: number;
}

export interface DailyPostCount {
  date: string;
  count: number;
}

export type GetListingsParams = {
  listing_type?: GetListingsListingType;
  property_type?: string;
  district?: string;
  phuket_zone?: GetListingsPhuketZone;
  poster_type?: GetListingsPosterType;
  opportunity_score?: GetListingsOpportunityScore;
  is_direct_owner?: boolean;
  has_pool?: boolean;
  min_price?: number;
  max_price?: number;
  min_bedrooms?: number;
  max_bedrooms?: number;
  limit?: number;
  offset?: number;
};

export type GetListingsListingType =
  (typeof GetListingsListingType)[keyof typeof GetListingsListingType];

export const GetListingsListingType = {
  for_sale: "for_sale",
  for_rent: "for_rent",
  unknown: "unknown",
} as const;

export type GetListingsPhuketZone =
  (typeof GetListingsPhuketZone)[keyof typeof GetListingsPhuketZone];

export const GetListingsPhuketZone = {
  north: "north",
  central: "central",
  south: "south",
  east: "east",
} as const;

export type GetListingsPosterType =
  (typeof GetListingsPosterType)[keyof typeof GetListingsPosterType];

export const GetListingsPosterType = {
  owner: "owner",
  agent: "agent",
  unknown: "unknown",
} as const;

export type GetListingsOpportunityScore =
  (typeof GetListingsOpportunityScore)[keyof typeof GetListingsOpportunityScore];

export const GetListingsOpportunityScore = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type GetDirectOwnersParams = {
  phuket_zone?: string;
  property_type?: string;
  min_price?: number;
  max_price?: number;
  has_pool?: boolean;
  date_from?: string;
  date_to?: string;
};

export type GetBuyerRequestsParams = {
  phuket_zone?: string;
  property_type?: string;
  date_from?: string;
  date_to?: string;
};