# lib/api-zod/src/generated/types/districtStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhuketDeal Real Estate Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface DistrictStat {
  district: string;
  count: number;
}