# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhuketDeal Real Estate Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  post_id: string;
  post_url?: string | null;
  author_name?: string | null;
  author_profile_url?: string | null;
  post_text?: string | null;
  created_at?: string | null;
  comment_count?: number | null;
  like_count?: number | null;
  screenshot_path?: string | null;
  classification_label?: string | null;
  is_direct_owner?: boolean | null;
  is_agent?: boolean | null;
  is_hot?: boolean | null;
  source_mode?: string | null;
  listing_type?: string | null;
  price_thb?: number | null;
  price_thb_min?: number | null;
  price_thb_max?: number | null;
  price_unit?: string | null;
  price_negotiable?: boolean | null;
  property_type?: string | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  floors?: number | null;
  size_sqm?: number | null;
  land_rai?: number | null;
  location?: string | null;
  district?: string | null;
  phuket_zone?: string | null;
  project_name?: string | null;
  condition?: string | null;
  furnished?: string | null;
  has_pool?: boolean | null;
  has_sea_view?: boolean | null;
  has_garden?: boolean | null;
  has_parking?: boolean | null;
  has_gym?: boolean | null;
  has_security?: boolean | null;
  has_mountain_view?: boolean | null;
  poster_type?: string | null;
  urgency_signals?: string[] | null;
  key_features?: string[] | null;
  decision_summary?: string | null;
  opportunity_score?: string | null;
  confidence?: string | null;
}