# lib/api-zod/src/generated/types/pricePerSqmStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhuketDeal Real Estate Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePerSqmStat {
  district: string;
  avg_price_per_sqm?: number | null;
  median_price_per_sqm?: number | null;
  min_price_per_sqm?: number | null;
  max_price_per_sqm?: number | null;
  count: number;
}