# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhuketDeal Real Estate Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  total_posts_this_week: number;
  total_posts_all_time: number;
  direct_owner_count: number;
  direct_owner_pct: number;
  for_sale_count: number;
  for_rent_count: number;
  high_opportunity_count: number;
  avg_price_thb?: number | null;
}